+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Change categories:
 - firmware update: Add support for PLDM 1.2 by adding `component_opaque_data`
   fields to correctly parse their contents.
 
+- Added the functionality for `pldm_edac_crc32_validate`.
+
 ### Changed
 
 - Stabilised `decode_get_event_receiver_resp()`